fix(StatsDao::getCountInDateRange()): returning values too low
Previously, the function would only return the time on one of the days, rather than the sum on all. This would not be a problem if the start and the end of the range were the same, but if the range would span over multiple days, the results would be incorrect.