fix: correctly manage the lifetime of BackupComponent
BackupComponent must be provided as a @Singleton because the lifetime needs to be as close to that of SingletonComponent as possible because BackupComponent's dependency subgraph contains dependencies that should act like singletons, i.e. BackupRepositoryImpl and RestoreBackupUseCaseImpl (prior to the introduction of the subcomponent, the 2 used to be annotated with @Singleton). The next step is to introduce a scope for BackupComponent and assign the singleton-like dependencies to that scope.