fix: correctly configure lifetimes of BackupRepositoryImpl and RestoreBackupUseCaseImpl
Instances of the 2 classes will only be created once per an instance of BackupComponent. Per application, there is a single instance of @Local BackupComponent and a single instance of @GoogleDrive BackupComponent. This means that per application, there are only 2 instances of BackupRepositoryImpl and RestoreBackupUseCaseImpl each - a @Local instance as part of the @Local BackupComponent and a @GoogleDrive instance as part of the @GoogleDrive BackupComponent.