refactor: move backup workers to the :data module
The only reason BackupWorker and its implementations previously found themselves in the :app module is that so did BackupComponent, which they depend on. Now that the component is back in :data, there is no reason for the workers to stay in the :app module in which they do not belong.