fix(TheApplication): ensure the new GoogleDriveBackupWorker replaces legacy BackupWorker
This is achieved by changing the new worker's tag to differ from that of BackupWorker. This way, when an attempt at invoking the legacy worker is performed, an exception will be thrown and caught, and as a result, the work request will be removed from the queue. The new worker, in the meantime, can be enqueued separately.